requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cloudinary==1.31.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
defusedxml==0.7.1
dj-database-url==1.2.0
dj-rest-auth==2.2.7
Django==4.1.5
django-allauth==0.51.0
django-cors-headers==3.13.0
django-enumfields==2.1.1
django-environ==0.9.0
django-jazzmin==2.6.0
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
